Changes to be committed: 	modified:   start.sh
Reworked the file so that it checks if it can authenticate on the deluge Web UI before it does anything else.
I did this so that the Port would not be instantly set when you aren't even able to set it which would result in the file waiting for the next port change.